Move api calls to TDK.API class
