+++ Assets/Treasure/TDK/Runtime/API/Magicswap.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 30a332c16ce884465af4d3c4db60b057
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
