+++ Assets/Treasure/TDK/Runtime/Magicswap/TDK.Magicswap.cs
@@ -1,8 +1,5 @@
-using System;
 using System.Collections.Generic;
-using System.Numerics;
 using System.Threading.Tasks;
-using Newtonsoft.Json;
 using UnityEngine;
 
 namespace Treasure
@@ -20,146 +17,20 @@ namespace Treasure
     public partial class Magicswap
     {
         public Magicswap() { }
-
-        [Serializable]
-        public class MagicswapToken {
-            [Serializable]
-            public class VaultCollection {
-                [Serializable]
-                public class VaultCollectionData {
-                    public string id;
-                    public string type;
-                }
-
-                public VaultCollectionData collection;
-                public List<string> tokenIds;
-            }
-
-            [Serializable]
-            public class Collection {
-                public string id;
-                public string urlSlug;
-                public List<string> tokenIds;
-                public string name;
-                public string symbol;
-                public string type;
-                public string image;
-            }
-
-            public string id;
-            public string name;
-            public string symbol;
-            public int decimals;
-            public string derivedMAGIC;
-            public bool isNFT;
-            public List<VaultCollection> vaultCollections;
-            public string type;
-            public string image;
-            public bool isMAGIC;
-            public List<Collection> collections;
-            public string urlSlug;
-            public string collectionId;
-            public List<string> collectionTokenIds;
-            public int priceUSD; // TODO check type; big, int or double? examples are 0
-            public string reserve;
-        }
-
         
-        
-        [Serializable]
-        public class MagicswapPool {
-            [Serializable]
-            public class DayData {
-                public string reserveUSD;
-                public string volumeUSD;
-                public string txCount;
-            }
-
-            public string id;
-            public MagicswapToken token0;
-            public MagicswapToken token1;
-            public string reserve0;
-            public string reserve1;
-            public BigInteger reserveUSD; // TODO check type; big, int or double? examples are 0
-            public string totalSupply;
-            public string txCount;
-            public double volumeUSD;
-            public string lpFee;
-            public string protocolFee;
-            public string royaltiesFee;
-            public string royaltiesBeneficiary = null;
-            public string totalFee;
-            public List<DayData> dayData; // TODO confirm naming
-            public string name;
-            public bool hasNFT;
-            public bool isNFTNFT;
-            public double volume24h0;
-            public double volume1w0;
-            public double volume24h1;
-            public double volume1w1;
-        }
-
-        [Serializable]
-        public class MagicswapRoute {
-            [Serializable]
-            public class RouteLeg {
-                public class LegTokenInfo {
-                    public string name;
-                    public string symbol;
-                    public string address;
-                    public int decimals;
-                    public string tokenId;
-                }
-
-                public string poolAddress;
-                public string poolType;
-                public double poolFee;
-                public LegTokenInfo tokenFrom;
-                public LegTokenInfo tokenTo;
-                public BigInteger assumedAmountIn;
-                public BigInteger assumedAmountOut;
-                public int swapPortion; // TODO check type; examples are 1
-                public int absolutePortion; // TODO check type; examples are 1
-            }
-
-            public string amountIn;
-            public string amountOut;
-            public MagicswapToken tokenIn;
-            public MagicswapToken tokenOut;
-            public List<RouteLeg> legs;
-            public List<string> path;
-            public double priceImpact;
-            public int derivedValue; // TODO check type; examples are small-ish ints
-            public double lpFee;
-            public double protocolFee; // TODO check type; examples are 0
-            public double royaltiesFee; // TODO check type; examples are 0
-        }
-
-        [Serializable]
-        public class MagicswapPoolsResponse {
-            public List<MagicswapPool> pools;
-        }
-        
-        // TODO should these move to API folder? (like Harvester.cs) it would be misdirection, but also more consistent
         public async Task<List<MagicswapPool>> GetAllPools()
         {
-            var response = await TDK.API.Get("/magicswap/pools");
-            return JsonConvert.DeserializeObject<MagicswapPoolsResponse>(response).pools;
+            return await TDK.API.GetAllPools();
         }
         
         public async Task<MagicswapPool> GetPoolById(string id)
         {
-            var response = await TDK.API.Get($"/magicswap/pools/{id}");
-            return JsonConvert.DeserializeObject<MagicswapPool>(response);
+            return await TDK.API.GetPoolById(id);
         }
 
         public async Task<MagicswapRoute> GetRoute(string tokenInId, string tokenOutId, string amount, bool isExactOut)
         {
-            var body = JsonConvert.SerializeObject(new { tokenInId, tokenOutId, amount, isExactOut });
-            var response = await TDK.API.Post("/magicswap/route", body);
-            return JsonConvert.DeserializeObject<MagicswapRoute>(response, new JsonSerializerSettings() {
-                MissingMemberHandling = MissingMemberHandling.Error // TODO remove this after further testing
-            });
+            return await TDK.API.GetRoute(tokenInId, tokenOutId, amount, isExactOut);
         }
     }
 }
